requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Deprecated==1.3.1
Django==6.0
djangorestframework==3.16.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-tools==1.71.2
idna==3.11
protobuf==5.29.5
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
pytest==8.0.0
pytest-django==4.8.0
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wrapt==2.0.1
yandexcloud==0.373.0